refactor: simplify frontend by extracting shared sidebar logic and removing dead code
Extract useSidebarData hook to deduplicate ~150 lines of identical data
logic between Sidebar and MobileSidebar. Add invalidateFeedDependents
helper to DRY up repeated query invalidation across feed mutations.
Fix longPressTimer useState→useRef to avoid unnecessary re-renders on
touch. Remove dead ARTICLE_LIST_PAGE_SIZE constant, trivial
shouldShowFolderUnreadBadge wrapper, duplicate Escape key handler, and
unused parentWeight prop.
