+++ frontend/src/components/feed/FeedRow.tsx
@@ -1,6 +1,6 @@
 "use client";
 
-import { useState, useCallback } from "react";
+import { useState, useCallback, useRef } from "react";
 import { Box, Flex, Text, Badge, IconButton, Input } from "@chakra-ui/react";
 import { LuCheckCheck, LuFolderInput, LuGripVertical, LuTrash2 } from "react-icons/lu";
 import { useSortable } from "@dnd-kit/sortable";
@@ -36,7 +36,7 @@ export function FeedRow({
 }: FeedRowProps) {
   const [isHovered, setIsHovered] = useState(false);
   const [isRevealed, setIsRevealed] = useState(false);
-  const [longPressTimer, setLongPressTimer] = useState<NodeJS.Timeout | null>(null);
+  const longPressTimer = useRef<ReturnType<typeof setTimeout> | null>(null);
 
   const handleRename = useCallback(
     (newName: string) => onRename(feed.id, newName),
@@ -77,23 +77,22 @@ export function FeedRow({
 
   // Mobile long-press for rename
   const handleTouchStart = () => {
-    const timer = setTimeout(() => {
+    longPressTimer.current = setTimeout(() => {
       startRename();
     }, 500);
-    setLongPressTimer(timer);
   };
 
   const handleTouchEnd = () => {
-    if (longPressTimer) {
-      clearTimeout(longPressTimer);
-      setLongPressTimer(null);
+    if (longPressTimer.current) {
+      clearTimeout(longPressTimer.current);
+      longPressTimer.current = null;
     }
   };
 
   const handleTouchMove = () => {
-    if (longPressTimer) {
-      clearTimeout(longPressTimer);
-      setLongPressTimer(null);
+    if (longPressTimer.current) {
+      clearTimeout(longPressTimer.current);
+      longPressTimer.current = null;
     }
   };
